requirements.txt
# OtterVision Dependencies
#
# STEP 1: Install PyTorch with CUDA 12.1 FIRST:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# STEP 2: Install everything else:
#   pip install -r requirements.txt

# Core
numpy>=1.24.0,<2.0.0

# Transformers (GroundingDINO + SAM for data prep, LLaVA for Q&A)
transformers>=4.37.0
accelerate>=0.25.0

# YOLOv8 (fast real-time detection + segmentation)
ultralytics>=8.1.0

# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0

# Video processing
opencv-python>=4.8.0
Pillow>=10.0.0

# YouTube download + splitting
yt-dlp>=2023.10.0
# Also requires: sudo apt install ffmpeg

# Training
scikit-learn>=1.3.0
matplotlib>=3.7.0
